split file_info.rs from config.rs.
